+++ fsl_mrs/dynamic/variable_mapping.py
@@ -104,6 +104,18 @@ class VariableMapping(object):
                         f' (parameter: {mp_obj.name}) not found in config file.')
                 _ = self.get_gradient_fcn(mp_obj)
 
+        # Check that all custom _init functions have an equivalent dynamic function
+        # and that they will be called by one or more parameters
+        for key in self.fcns:
+            if '_init' in key:
+                expected_dyn = key.replace('_init', '')
+                if expected_dyn not in self.fcns:
+                    raise ConfigFileError(
+                        f'Custom init function {key} does not have matching dynamic function {expected_dyn}.')
+                if expected_dyn not in self.free_functions:
+                    raise ConfigFileError(
+                        f'Custom init function {key} will not be called as {expected_dyn} is not used.')
+
     def __str__(self):
         OUT  = '-----------------------\n'
         OUT += 'Variable Mapping Object\n'
@@ -180,11 +192,17 @@ class VariableMapping(object):
 
         def grad_function(self):
             """Return gradient function name"""
-            return self.function_name + '_grad'
+            if self.param_type == 'dynamic':
+                return self.function_name + '_grad'
+            else:
+                return None
 
         def init_function(self):
             """Return initialisation function name"""
-            return self.function_name + '_init'
+            if self.param_type == 'dynamic':
+                return self.function_name + '_init'
+            else:
+                return None
 
     @property
     def mapped_parameters(self):
@@ -452,7 +470,7 @@ class VariableMapping(object):
         """
         Convert free into mapped params over time
         fixed params get copied over time domain
-        variable params are indep over time
+        variable params are independent over time
         dynamic params are mapped using dyn model
 
         Parameters
@@ -516,8 +534,10 @@ class VariableMapping(object):
     def mapped_to_free(self, p):
         """
         Convert mapped params to free (e.g. to initialise the free params)
-        fixed and variable params are simply copied
-        dynamic params are converted by inverting the dyn model with Scipy optimize
+        Fixed free parameters are initialised as the median of the individual values
+        Variable free params are simply copied
+        Dynamic free params are calculated by either using a defined custom initialisation (..._init)
+        or through the default init function that tries to invert the dynamic model with Scipy optimize.
 
         Parameters
         ----------
@@ -538,7 +558,6 @@ class VariableMapping(object):
         free_params = np.zeros(self.nfree)
         for index, mp_obj in enumerate(self._mapped_params):
             init_function = self.get_init_fcn(mp_obj)
-
             free_params[mp_obj.free_indices] = init_function(p[:, index], self.time_variable)
 
         return free_params